refactor : SyncRepository 추가
