+++ core/data/src/main/java/com/tgyuu/data/repository/ConfigRepositoryImpl.kt
@@ -1,6 +1,6 @@
 package com.tgyuu.data.repository
 
-import com.tgyuu.datastore.datasource.LocalUserConfigDataSource
+import com.tgyuu.datastore.datasource.user.LocalUserConfigDataSource
 import com.tgyuu.domain.model.SortType
 import com.tgyuu.domain.model.UpdateInfo
 import com.tgyuu.domain.repository.ConfigRepository
@@ -39,7 +39,4 @@ class ConfigRepositoryImpl @Inject constructor(
             defaultValue = GetUpdateInfoResponse(),
         ).toDomain()
     }
-
-    override suspend fun ensureUUIDExists() = localUserConfigDataSource.ensureUUIDExists()
-    override suspend fun getUUID(): String = localUserConfigDataSource.uuid.first()
 }
