+++ core/datastore/src/main/java/com/tgyuu/datastore/datasource/sync/LocalSyncDataSource.kt
@@ -1,0 +1,9 @@
+package com.tgyuu.datastore.datasource.sync
+
+import kotlinx.coroutines.flow.Flow
+
+interface LocalSyncDataSource {
+    val uuid: Flow<String>
+    suspend fun ensureUUIDExists()
+    suspend fun setUuid(uuid: String)
+}
