+++ core/datastore/src/main/java/com/tgyuu/datastore/datasource/user/LocalUserConfigDataSource.kt
@@ -1,4 +1,4 @@
-package com.tgyuu.datastore.datasource
+package com.tgyuu.datastore.datasource.user
 
 import com.tgyuu.domain.model.SortType
 import kotlinx.coroutines.flow.Flow
@@ -7,11 +7,8 @@ interface LocalUserConfigDataSource {
     val sortType: Flow<SortType>
     val notificationEnabled: Flow<Boolean>
     val alarmTime: Flow<Pair<Int, Int>>
-    val uuid: Flow<String>
     suspend fun consumeIsFirstAppOpen(): Boolean
     suspend fun setSortType(sortType: SortType)
     suspend fun setNotificationEnabled(enabled: Boolean)
     suspend fun setAlarmTime(hour: String, minute: String)
-    suspend fun ensureUUIDExists()
-    suspend fun setUuid(uuid: String)
 }
