+++ core/datastore/src/main/java/com/tgyuu/datastore/di/DataStoreProvidesModule.kt
@@ -4,8 +4,8 @@ import android.content.Context
 import androidx.datastore.core.DataStore
 import androidx.datastore.preferences.core.Preferences
 import androidx.datastore.preferences.preferencesDataStore
-import com.tgyuu.datastore.datasource.LocalUserConfigDataSource
-import com.tgyuu.datastore.datasource.LocalUserConfigDataSourceImpl
+import com.tgyuu.datastore.datasource.user.LocalUserConfigDataSource
+import com.tgyuu.datastore.datasource.user.LocalUserConfigDataSourceImpl
 import dagger.Binds
 import dagger.Module
 import dagger.Provides
@@ -21,12 +21,22 @@ object DataStoreProvidesModule {
     private const val CONFIG_DATASTORE_NAME = "CONFIGS_PREFERENCES"
     private val Context.configDataStore by preferencesDataStore(name = CONFIG_DATASTORE_NAME)
 
+    private const val SYNC_DATASTORE_NAME = "SYNC_PREFERENCES"
+    private val Context.syncDataStore by preferencesDataStore(name = SYNC_DATASTORE_NAME)
+
     @Provides
     @Singleton
     @Named("config")
     fun provideConfigDataStore(
         @ApplicationContext context: Context
     ): DataStore<Preferences> = context.configDataStore
+
+    @Provides
+    @Singleton
+    @Named("sync")
+    fun provideSyncDataStore(
+        @ApplicationContext context: Context
+    ): DataStore<Preferences> = context.syncDataStore
 }
 
 @Module
