+++ core/domain/src/main/java/com/tgyuu/domain/repository/SyncRepository.kt
@@ -1,0 +1,6 @@
+package com.tgyuu.domain.repository
+
+interface SyncRepository {
+    suspend fun ensureUUIDExists()
+    suspend fun getUUID(): String
+}
