+++ feature/sync/src/main/java/com/tgyuu/sync/graph/download/DownloadViewModel.kt
@@ -3,7 +3,7 @@ package com.tgyuu.sync.graph.download
 import androidx.lifecycle.viewModelScope
 import com.tgyuu.common.base.BaseViewModel
 import com.tgyuu.common.event.EventBus
-import com.tgyuu.domain.repository.ConfigRepository
+import com.tgyuu.domain.repository.SyncRepository
 import com.tgyuu.navigation.NavigationBus
 import com.tgyuu.navigation.NavigationEvent
 import com.tgyuu.sync.graph.download.contract.DownloadIntent
@@ -14,14 +14,14 @@ import javax.inject.Inject
 
 @HiltViewModel
 class DownloadViewModel @Inject constructor(
-    private val configRepository: ConfigRepository,
+    private val syncRepository: SyncRepository,
     private val navigationBus: NavigationBus,
     private val eventBus: EventBus,
 ) : BaseViewModel<DownloadState, DownloadIntent>(DownloadState()) {
 
     init {
         viewModelScope.launch {
-            val uuid = configRepository.getUUID()
+            val uuid = syncRepository.getUUID()
             setState { copy(uuid = uuid) }
         }
     }
