+++ feature/sync/src/main/java/com/tgyuu/sync/graph/link/LinkViewModel.kt
@@ -3,7 +3,7 @@ package com.tgyuu.sync.graph.link
 import androidx.lifecycle.viewModelScope
 import com.tgyuu.common.base.BaseViewModel
 import com.tgyuu.common.event.EventBus
-import com.tgyuu.domain.repository.ConfigRepository
+import com.tgyuu.domain.repository.SyncRepository
 import com.tgyuu.navigation.NavigationBus
 import com.tgyuu.navigation.NavigationEvent
 import com.tgyuu.sync.graph.link.contract.LinkIntent
@@ -14,14 +14,14 @@ import javax.inject.Inject
 
 @HiltViewModel
 class LinkViewModel @Inject constructor(
-    private val configRepository: ConfigRepository,
+    private val syncRepository: SyncRepository,
     private val navigationBus: NavigationBus,
     private val eventBus: EventBus,
 ) : BaseViewModel<LinkState, LinkIntent>(LinkState()) {
 
     init {
         viewModelScope.launch {
-            val uuid = configRepository.getUUID()
+            val uuid = syncRepository.getUUID()
             setState { copy(uuid = uuid) }
         }
     }
