+++ feature/sync/src/main/java/com/tgyuu/sync/graph/upload/contract/UploadIntent.kt
@@ -4,5 +4,6 @@ import com.tgyuu.common.base.UiIntent
 
 sealed interface UploadIntent : UiIntent {
     data object OnBackClick : UploadIntent
+    data object OnUuidClick : UploadIntent
     data object OnUploadClick : UploadIntent
 }
